backend/requirements.txt
# backend/requirements.txt

uuid
fastapi
fastai
numpy
pillow
python-multipart
uvicorn
